requirements.txt
SQLAlchemy
psycopg2-binary
langchain
langchain-community
langchain-openai
langchain-core
pandas
streamlit
python-dotenv
faiss-cpu
matplotlib
openai
